+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To be Released
 
+* Bump github.com/gofrs/uuid from 4.1.0+incompatible to 4.2.0+incompatible
+
 ## v1.4.1
 
 * Bump github.com/sirupsen/logrus from 1.7.0 to 1.8.1